Conflicts:
	CRM/Contact/BAO/Contact.php
	CRM/Group/Page/AJAX.php
	CRM/Report/Form/Member/Detail.php
	CRM/Utils/SoapServer.php
	CRM/Utils/System/Base.php
	sql/civicrm_generated.mysql
	tests/phpunit/CRM/Core/DAOTest.php
	tests/phpunit/CiviTest/CiviUnitTestCase.php
	tests/phpunit/api/v3/ContributionTest.php
	xml/version.xml

+++ CRM/Case/XMLProcessor/Process.php
@@ -120,7 +120,8 @@ class CRM_Case_XMLProcessor_Process extends CRM_Case_XMLProcessor {
   }
 
   function processStandardTimeline($activitySetXML, &$params) {
-    if ('Change Case Type' == CRM_Utils_Array::value('activityTypeName', $params)) {
+    if ('Change Case Type' == CRM_Utils_Array::value('activityTypeName', $params)
+      && CRM_Utils_Array::value('resetTimeline', $params, TRUE)) {
       // delete all existing activities which are non-empty
       $this->deleteEmptyActivity($params);
     }
